requirements.txt
azure-core==1.25.0
azure-eventhub==5.10.0
azure-iot-device==2.11.0
azure-iot-hub==2.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
deprecation==2.1.0
et-xmlfile==1.1.0
idna==3.3
isodate==0.6.1
janus==1.0.0
msrest==0.7.1
numpy==1.23.2
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
pip==22.2.2
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
setuptools==58.1.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.3.0
uamqp==1.5.3
urllib3==1.26.11
